test: fix act warnings in test
